% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{example}
\alias{example}
\title{Example data frame data}
\usage{
data(example)
}
\value{
A data frame
}
\description{
Used only for the introduction to data frames example
}
\examples{
data(example)
}
\keyword{datasets}
