% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_path.R
\name{get_got_path}
\alias{get_got_path}
\title{Game of Thrones CSV Path}
\usage{
get_got_path()
}
\description{
Returns the path to the Game of thrones data set.
}
\examples{
fname = get_got_path()
readr::read_csv(fname)
}
