% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes.R
\name{get_notes}
\alias{get_notes}
\title{View the notes for the jrTidyverse course}
\usage{
get_notes()
}
\description{
This function opens a browser to show the pdf of the
 handout notes. This way attendees can always get those notes back.
}
