% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{GoT}
\alias{GoT}
\alias{GoT_df}
\title{Game of Thrones Audience figures}
\value{
A tibble or data frame.
}
\description{
Taken from wikipedia. The data set \code{GoT_df} returns a
data frame, whereas \code{GoT} returns a tibble.
}
\examples{
data(GoT)
data(GoT_df)
}
\keyword{datasets}
