% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{School}
\alias{School}
\title{School grade example}
\usage{
data(School)
}
\value{
A data frame
}
\description{
Used for spread and gather examples
}
\examples{
data(School)
}
\keyword{datasets}
