% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{stock_change}
\alias{stock_change}
\title{Stock price changes example}
\usage{
data(stock_change)
}
\value{
A data frame
}
\description{
Used only for the dplyr joining example
}
\examples{
data(stock_change)
}
\keyword{datasets}
