% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{stock_price}
\alias{stock_price}
\title{Stock prices example}
\usage{
data(stock_price)
}
\value{
A data frame
}
\description{
Used only for the dplyr joining example
}
\examples{
data(stock_price)
}
\keyword{datasets}
